+++ packages/liveblocks-core/src/index.ts
@@ -233,6 +233,7 @@ export type {
 } from "./protocol/Comments";
 export type { QueryMetadata } from "./protocol/Comments";
 export type {
+  SearchCommentsResult,
   ThreadData,
   ThreadDataPlain,
   ThreadDataWithDeleteInfo,