+++ packages/liveblocks-react/src/types/index.ts
@@ -42,6 +42,7 @@ import type {
   Relax,
   Resolve,
   RoomEventMessage,
+  SearchCommentsResult,
   SyncStatus,
   ThreadData,
   ToImmutable,
@@ -162,6 +163,33 @@ export type UseThreadsOptions<M extends BaseMetadata> = {
   scrollOnLoad?: boolean;
 };
 
+export type SearchCommentsQuery<M extends BaseMetadata> = {
+  /**
+   * (Optional) Metadata to filter the threads by.
+   */
+  threadMetadata?: Partial<QueryMetadata<M>>;
+  /**
+   * (Optional) Whether to only return comments from threads marked as resolved or unresolved.
+   */
+  threadResolved?: boolean;
+  /**
+   * (Optional) Whether to only return comments that have attachments.
+   */
+  hasAttachments?: boolean;
+  /**
+   * (Optional) Whether to only return comments that have mentions.
+   */
+  hasMentions?: boolean;
+  /**
+   * (Required) Text to search within comment content. Uses rich text and vector search for relevance.
+   */
+  text: string;
+};
+
+export type UseSearchCommentsOptions<M extends BaseMetadata> = {
+  query: SearchCommentsQuery<M>;
+};
+
 export type InboxNotificationsQuery = {
   /**
    * Whether to only return inbox notifications for a specific room.
@@ -254,6 +282,8 @@ export type PagedAsyncResult<T, F extends string> = Relax<
 
 export type ThreadsAsyncSuccess<M extends BaseMetadata> = PagedAsyncSuccess<ThreadData<M>[], "threads">; // prettier-ignore
 export type ThreadsAsyncResult<M extends BaseMetadata> = PagedAsyncResult<ThreadData<M>[], "threads">; // prettier-ignore
+
+export type SearchCommentsAsyncResult = AsyncResult<Array<SearchCommentsResult>, "results">; // prettier-ignore
 
 export type InboxNotificationsAsyncSuccess = PagedAsyncSuccess<InboxNotificationData[], "inboxNotifications">; // prettier-ignore
 export type InboxNotificationsAsyncResult = PagedAsyncResult<InboxNotificationData[], "inboxNotifications">; // prettier-ignore
@@ -1097,6 +1127,16 @@ export type RoomContextBundle<
        * const { threads, error, isLoading } = useThreads();
        */
       useThreads(options?: UseThreadsOptions<M>): ThreadsAsyncResult<M>;
+
+      /**
+       * Returns the result of searching comments by text in the current room. The result includes the id and the plain text content of the matched comments along with the parent thread id of the comment.
+       *
+       * @example
+       * const { results, error, isLoading } = useSearchComments({ query: { text: "hello"} });
+       */
+      useSearchComments(
+        options: UseSearchCommentsOptions<M>
+      ): SearchCommentsAsyncResult;
 
       /**
        * Returns the user's subscription settings for the current room